==> BreCaNet/BreCaNet/1_SubtypeLabels.Rmd <==
---
title: "fromBiostars"
author: "Ursula Widocki"
date: "3/1/2020"
output: html_document
---

```{r setup, include=FALSE}
#knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())
```

```{r}
library(dplyr)
```

```{r}
brca_subtypes <- TCGAbiolinks::TCGAquery_subtype("brca")
colnames(brca_subtypes) # need BRCA_Subtype_PAM50
```

```{r}
# makes a file of just the information I need
brca_sample_types <- cbind(data.frame(brca_subtypes$patient), data.frame(brca_subtypes$BRCA_Subtype_PAM50))

write.table(brca_sample_types, file = "TCGA_subtypes.txt", sep = " ", row.names = TRUE, col.names = TRUE)

```

```{r}
table(brca_sample_types$brca_subtypes.BRCA_Subtype_PAM50)

```
