==> BreCaNet/BreCaNet/2_RNAseq_for_networks.ipynb <==
# BreCaNet

This Jupyter notebook goes through the steps to process the full data sets before running PANDA

##### Gene Expression Data

HTSeq
FPKM
metadata

import numpy as np
import os
import shutil
import csv
import pandas as pd
import gzip
import json

## Paths to expression data

dataPath_HTSeq = '/Users/ursulawidocki/Desktop/BreCaNet/Data/gdc_download_HTSeq' #zipped file path
dataNewFolder_HTSeq = '/Users/ursulawidocki/Desktop/BreCaNet/Data/processedTCGA_HTSeq'

dataPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/gdc_download_FPKM_832' #zipped file path
dataNewFolder = '/Users/ursulawidocki/Desktop/BreCaNet/Data/processedTCGA_832' # folder for unzipped

## Unzips HTSeq expression data
# creates a folder (processedTCGA_HTSeq) of the unzipped files 
# Gets the codes of the patient folders and expression files

folderIDs_HTSeq = list() # codes of the patient IDs on the folder
indivList_HTSeq = list() # codes of the patients' expression file

for fileName in os.listdir(dataPath_HTSeq):
    if fileName == '.DS_Store' or fileName == 'MANIFEST.txt':
        continue
        
    else:
        folderIDs_HTSeq.append(fileName)
        
    newPath = dataPath_HTSeq + '/' + fileName
    
    # goes into the individual folder
    for indivName in os.listdir(newPath):
        if (indivName == '.DS_Store') or (indivName == 'MANIFEST.txt') or (indivName == 'annotations.txt'):
            continue
        filePath = newPath + '/' + indivName
        
        fileEnding = indivName[-3] + indivName[-2] + indivName[-1]
        if(fileEnding == '.gz'):
            with gzip.open(filePath, 'rb') as f_in:
                newFilePath = dataNewFolder_HTSeq + '/' + indivName[0:-3]
                indivList_HTSeq.append(indivName[0:-3])
                
                with open(newFilePath, 'wb') as f_out:
                    shutil.copyfileobj(f_in, f_out)
                    
                    
## Unzips FPKM expression data
# creates a folder (processedTCGA_FPKM) of the unzipped files 
# Gets the codes of the patient folders and expression files

folderIDs = list() # codes of the patient IDs on the folder
indivList = list() # codes of the patients' expression file

for fileName in os.listdir(dataPath):
    if fileName == '.DS_Store' or fileName == 'MANIFEST.txt':
        continue
        
    else:
        folderIDs.append(fileName)
        
    newPath = dataPath + '/' + fileName
    
    # goes into the individual folder
    for indivName in os.listdir(newPath):
        if (indivName == '.DS_Store') or (indivName == 'MANIFEST.txt') or (indivName == 'annotations.txt'):
            continue
        filePath = newPath + '/' + indivName
        
        fileEnding = indivName[-3] + indivName[-2] + indivName[-1]
        if(fileEnding == '.gz'):
            with gzip.open(filePath, 'rb') as f_in:
                newFilePath = dataNewFolder + '/' + indivName[0:-3]
                indivList.append(indivName[0:-3])
                
                with open(newFilePath, 'wb') as f_out:
                    shutil.copyfileobj(f_in, f_out)

## Checks if folder IDs and file IDs are similar between the two 
# expression sets

print(len(folderIDs_HTSeq) == len(folderIDs))
print(len(indivList_HTSeq) == len(indivList))

## Makes file of folder names

folderIDPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/foldIDList_HTSeq.txt'
folderFile = open(folderIDPath, "w+")
for i in range(0, 832):
    newLineFolder = folderIDs_HTSeq[i] + "\n"
    
    folderFile.write(newLineFolder)
    
folderFile.close()

folderIDPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/foldIDList_FPKM_832.txt'
folderFile = open(folderIDPath, "w+")
for i in range(0, 832):
    newLineFolder = folderIDs[i] + "\n"
    
    folderFile.write(newLineFolder)
    
folderFile.close()

## Gets folder codes from expression data

folderIDPath_HTSeq = '/Users/ursulawidocki/Desktop/BreCaNet/Data/foldIDList_HTSeq.txt'
folders_HTSeq = pd.read_csv(folderIDPath_HTSeq, delimiter = '\t', header)
folders_list_HTSeq = list(folders_HTSeq[0])

folderIDPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/foldIDList_FPKM_832.txt'
folders = pd.read_csv(folderIDPath, delimiter = '\t', header = None)
folders_list = list(folders[0])

## Reads in metadata
# reads in case_ids such that only the first instance is used

# HTSeq
# gets file_id_HTSeq and case_id_HTSeq
metaPath_HTSeq = '/Users/ursulawidocki/Desktop/BreCaNet/Data/metadata.cart.HTSeq.json'
with open(metaPath_HTSeq, 'r') as f:
    meta_dict_HTSeq = json.load(f)

file_id_HTSeq = list()
case_id_HTSeq = list()
for i in range(0, len(meta_dict_HTSeq)):
    if(meta_dict_HTSeq[i].get("associated_entities")[0].get("case_id") not in case_id_HTSeq):
        file_id_HTSeq.append(meta_dict_HTSeq[i].get("file_id"))
        case_id_HTSeq.append(meta_dict_HTSeq[i].get("associated_entities")[0].get("case_id"))
            

print(len(file_id_HTSeq))
print(len(case_id_HTSeq))
print(len(set(file_id_HTSeq)))
print(len(set(case_id_HTSeq)))

# FPKM
# gets file_id and case_id
metaPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/metadata.cart.FPKM_832.json'
with open(metaPath, 'r') as f:
    meta_dict = json.load(f)
    
file_id = list()
case_id = list()
for i in range(0, len(meta_dict)):
    if(meta_dict[i].get("associated_entities")[0].get("case_id") not in case_id):
        file_id.append(meta_dict[i].get("file_id"))
        case_id.append(meta_dict[i].get("associated_entities")[0].get("case_id"))

print(len(file_id))
print(len(case_id))
print(len(set(file_id)))
print(len(set(case_id)))

print(file_id_HTSeq == file_id)
print(set(file_id_HTSeq) == set(file_id))
print(case_id_HTSeq == case_id)
print(set(case_id_HTSeq) == set(case_id)) # since True, these two datasets are of same samples

## Gets uuid and barcodes of the patients from clinical data

# reads in clinical information to link patients with their barcode
clinicalPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/gdc_download_clinical_832/8162d394-8b64-4da2-9f5b-d164c54b9608/nationwidechildrens.org_clinical_patient_brca.txt'
clinical_df = pd.read_csv(clinicalPath, delimiter = '\t')

# removes rows that are not necessary to match data
clinical_df = clinical_df.drop(clinical_df.index[0])
clinical_df = clinical_df.drop(clinical_df.index[0])
clinical_df.head()
codes_df = clinical_df[['bcr_patient_uuid', 'bcr_patient_barcode']]
#codes_df.head()

uuid_list = list(codes_df['bcr_patient_uuid'])
barcode_list = list(codes_df['bcr_patient_barcode'])

print(len(uuid_list))
print(len(barcode_list))

## Reads in PAM50 labels

subtype_path = "/Users/ursulawidocki/Desktop/BreCaNet/Code/TCGA_subtypes.txt"
subtypes = pd.read_csv(subtype_path, delimiter = " ")
subtypes.head()

pt_barcodes_list = list(subtypes['brca_subtypes.patient'])
subtype_list = list(subtypes['brca_subtypes.BRCA_Subtype_PAM50'])

print(len(pt_barcodes_list))
print(len(subtype_list))

### Matching

## First, match the folder with the file_ids to get the corresponding cases from the metadata

match_cases_HTSeq = list()

for folder in folders_list_HTSeq:
    if folder in file_id_HTSeq:
        ind = file_id_HTSeq.index(folder)
        match_cases_HTSeq.append(case_id_HTSeq[ind])


## FPKM

match_cases = list()

for folder in folders_list:
    if folder in file_id:
        ind = file_id.index(folder)
        match_cases.append(case_id[ind])

print(set(match_cases_HTSeq) == set(match_cases))
print(len(match_cases_HTSeq))
print(len(match_cases))

# there are 14 file_ids that cannot be matched to expression data

## Matches case_ids from metadata with uuid and patient barcodes from clinical data

match_barcodes_HTSeq = list()

for case in match_cases_HTSeq:
    for ptID in uuid_list:
        if case == ptID.lower():
            ind = uuid_list.index(ptID)
            match_barcodes_HTSeq.append(barcode_list[ind])
            break


match_barcodes = list()

for case in match_cases:
    for ptID in uuid_list:
        if case == ptID.lower():
            ind = uuid_list.index(ptID)
            match_barcodes.append(barcode_list[ind])
            break
            
print(set(match_barcodes_HTSeq) == set(match_barcodes))
print(len(match_barcodes_HTSeq))
print(len(match_barcodes))

## Matches subtypes to their barcodes from the TCGA data

# HTSeq
match_subtypes_HTSeq = list()

for barcode in match_barcodes_HTSeq:
    ind = pt_barcodes_list.index(barcode)
    match_subtypes_HTSeq.append(subtype_list[ind])


# FPKM
match_subtypes = list()

for barcode in match_barcodes:
    ind = pt_barcodes_list.index(barcode)
    match_subtypes.append(subtype_list[ind])

print(set(match_subtypes_HTSeq) == set(match_subtypes))
print(len(match_subtypes_HTSeq))
print(len(match_subtypes))

print(len(set(match_subtypes))) # there is only this many subtypes among my samples
print((set(match_subtypes)))

## gets list of genes that are in the file

# reads in HTSeq
geneList_HTSeq = list() # list of all genes that are noted in the expression files

with open(dataNewFolder_HTSeq + '/' + indivList_HTSeq[0], "r") as file:
    for line in file.readlines():
        potentialGene = line.strip().split("\t")[0]
        
        if potentialGene[0:3] == "ENS":
            newGeneName = potentialGene.strip().split(".")[0]
            geneList_HTSeq.append(newGeneName)
print(len(geneList_HTSeq))


# reads in FPKM
geneList = list() # list of all genes that are noted in the expression files

with open(dataNewFolder + '/' + indivList[0], "r") as file:
    for line in file.readlines():
        potentialGene = line.strip().split("\t")[0]
        
        if potentialGene[0:3] == "ENS":
            newGeneName = potentialGene.strip().split(".")[0]
            geneList.append(newGeneName)
print(len(geneList))

geneList_lumA = geneList
geneList_lumB = geneList
geneList_basal = geneList
geneList_her2 = geneList
geneList_norm = geneList

print(set(geneList_HTSeq) == set(geneList)) # note: genes not in same order in files but same genes present

## Processing raw counts (HTSeq) for DE analysis (done in R)


#### Now, sort the folder IDs into their subtypes
Notice that the order of folders_list_HTSeq is in the same order as match_subtypes_HTSeq

## uses index of subtype label to sort folder IDs into subtypes

# Luminal A samples
lumA_samples_HTSeq = list() # folder name of expression files
lumA_indiv_HTSeq = list() # labels of expression files
for i in range(0,len(match_subtypes_HTSeq)):
    if match_subtypes_HTSeq[i]=="LumA":
        lumA_samples_HTSeq.append(file_id_HTSeq[i])
        lumA_indiv_HTSeq.append(indivList_HTSeq[folderIDs_HTSeq.index(file_id_HTSeq[i])])
        
# Luminal B samples
lumB_samples_HTSeq = list()
lumB_indiv_HTSeq = list()
for i in range(0,len(match_subtypes_HTSeq)):
    if match_subtypes_HTSeq[i]=="LumB":
        lumB_samples_HTSeq.append(file_id_HTSeq[i])
        lumB_indiv_HTSeq.append(indivList_HTSeq[folderIDs_HTSeq.index(file_id_HTSeq[i])])
        
# Basal samples
basal_samples_HTSeq = list()
basal_indiv_HTSeq = list()
for i in range(0,len(match_subtypes_HTSeq)):
    if match_subtypes_HTSeq[i]=="Basal":
        basal_samples_HTSeq.append(file_id_HTSeq[i])
        basal_indiv_HTSeq.append(indivList_HTSeq[folderIDs_HTSeq.index(file_id_HTSeq[i])])
        
# Her2 samples
her2_samples_HTSeq = list()
her2_indiv_HTSeq = list()
for i in range(0,len(match_subtypes_HTSeq)):
    if match_subtypes[i]=="Her2":
        her2_samples_HTSeq.append(file_id_HTSeq[i])
        her2_indiv_HTSeq.append(indivList_HTSeq[folderIDs_HTSeq.index(file_id_HTSeq[i])])

# Normal samples
norm_samples_HTSeq = list()
norm_indiv_HTSeq = list()
for i in range(0,len(match_subtypes_HTSeq)):
    if match_subtypes_HTSeq[i]=="Normal":
        norm_samples_HTSeq.append(file_id_HTSeq[i])
        norm_indiv_HTSeq.append(indivList_HTSeq[folderIDs_HTSeq.index(file_id_HTSeq[i])])


## Makes a giant expression matrix for edgeR
# makes a matrix where patients are columns and genes are rows

indivList_HTSeq2 =lumA_indiv_HTSeq + lumB_indiv_HTSeq + basal_indiv_HTSeq + her2_indiv_HTSeq + norm_indiv_HTSeq
all_exp_HTSeq = np.zeros((len(geneList_HTSeq), len(indivList_HTSeq2))) # matrix with all info from subtypes
j = 0 # index of the column (sample)
i = 0 # index of the row (gene)
print(all_exp_HTSeq.shape) # just to make sure dimensions are correct

for fileName in os.listdir(dataNewFolder_HTSeq):
    if fileName == '.DS_Store' or fileName == 'MANIFEST.txt':
        continue
        
    if fileName in indivList_HTSeq2:
        temp_folder = file_id_HTSeq[indivList_HTSeq2.index(fileName)]
        readFile = dataNewFolder_HTSeq + '/' + fileName
    
        with open(readFile, "r") as file:
            for line in file.readlines():
                if("no_feature" in line):
                        break
                all_exp_HTSeq[i, j] = float(line.strip().split("\t")[1])
                i+=1
                
            i = 0
            if j != len(indivList_HTSeq2):
                j+=1

print(all_exp_HTSeq[0:5,])

## Makes a reference file of file names and their subtypes

ref_col_path = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/all_labels.txt'
refFile = open(ref_col_path,"w+")

for i in range(0, len(match_barcodes_HTSeq)):
    ref_row = match_barcodes_HTSeq[i] + "\t" + match_subtypes_HTSeq[i] + "\n"
    refFile.write(ref_row)
    
refFile.close()

## Processing FPKM data for PANDA networks


#### Now, sort the folder IDs into their subtypes
Notice that the order of folders_list is in the same order as match_subtypes

## uses index of subtype label to sort folder IDs into subtypes

# Luminal A samples
lumA_samples = list()
lumA_indiv = list()
for i in range(0,len(match_subtypes)):
    if match_subtypes[i]=="LumA":
        lumA_samples.append(folders_list[i])
        lumA_indiv.append(indivList[folderIDs.index(file_id[i])])

# Luminal B samples
lumB_samples = list()
lumB_indiv = list()
for i in range(0,len(match_subtypes)):
    if match_subtypes[i]=="LumB":
        lumB_samples.append(folders_list[i])
        lumB_indiv.append(indivList[folderIDs.index(file_id[i])])

# Basal samples
basal_samples = list()
basal_indiv = list()
for i in range(0,len(match_subtypes)):
    if match_subtypes[i]=="Basal":
        basal_samples.append(folders_list[i])
        basal_indiv.append(indivList[folderIDs.index(file_id[i])])

# Her2 samples
her2_samples = list()
her2_indiv = list()
for i in range(0,len(match_subtypes)):
    if match_subtypes[i]=="Her2":
        her2_samples.append(folders_list[i])
        her2_indiv.append(indivList[folderIDs.index(file_id[i])])

# Normal samples
norm_samples = list()
norm_indiv = list()
for i in range(0,len(match_subtypes)):
    if match_subtypes[i]=="Normal":
        norm_samples.append(folders_list[i])
        norm_indiv.append(indivList[folderIDs.index(file_id[i])])
        

## Reads through the unzipped expression files and creates the expression matrices depending on how the file
# name matches with the folder IDs of each subtype

lumA_exp = np.zeros((len(geneList), len(lumA_samples))) # exp matrix of LumA samples
lumB_exp = np.zeros((len(geneList), len(lumB_samples))) # exp matrix of LumB samples
basal_exp = np.zeros((len(geneList), len(basal_samples))) # exp matrix of basal samples
her2_exp = np.zeros((len(geneList), len(her2_samples))) # exp matrix of her2 samples
norm_exp = np.zeros((len(geneList), len(norm_samples))) # exp matrix of LumA samples
indivList2 = lumA_indiv + lumB_indiv + basal_indiv + her2_indiv + norm_indiv

print(norm_exp.shape) # just to make sure dimensions are correct
 
lumA_j = 0 # index of the column (sample)index of the column (sample)
lumB_j = 0
basal_j = 0
her2_j = 0
norm_j = 0
for fileName in os.listdir(dataNewFolder):
    if fileName == '.DS_Store' or fileName == 'MANIFEST.txt':
        continue
    
    if fileName in indivList2:
        
        temp_folder = folderIDs[indivList2.index(fileName)]
        readFile = dataNewFolder + '/' + fileName
            
        # adds to luminal A expression matrix
        if temp_folder in lumA_samples:
            
            lumA_i = 0
            with open(readFile, "r") as file:
                for line in file.readlines():
                    lumA_exp[lumA_i, lumA_j] = float(line.strip().split("\t")[1])
                    lumA_i += 1
                    
                lumA_j += 1
                        
        # adds to luminal B expression matrix
        if temp_folder in lumB_samples:
            
            lumB_i = 0
            with open(readFile, "r") as file:
                for line in file.readlines():
                    lumB_exp[lumB_i, lumB_j] = float(line.strip().split("\t")[1])
                    lumB_i += 1
                    
                lumB_j += 1
        
        # adds to basal expression matrix
        if temp_folder in basal_samples:
            
            basal_i = 0
            with open(readFile, "r") as file:
                for line in file.readlines():
                    basal_exp[basal_i, basal_j] = float(line.strip().split("\t")[1])
                    basal_i += 1
                    
                basal_j += 1
        
        # adds to her2 expression matrix
        if temp_folder in her2_samples:
            
            her2_i = 0
            with open(readFile, "r") as file:
                for line in file.readlines():
                    her2_exp[her2_i, her2_j] = float(line.strip().split("\t")[1])
                    her2_i += 1
                    
                her2_j += 1
        
        # adds to norm expression matrix
        if temp_folder in norm_samples:
            
            norm_i = 0
            with open(readFile, "r") as file:
                for line in file.readlines():
                    norm_exp[norm_i, norm_j] = float(line.strip().split("\t")[1])
                    norm_i += 1
                    
                norm_j += 1

## Removes zero rows from expression data

# Luminal A data
rowList_lumA = list()
for i in range(0, len(lumA_exp)):
    if np.sum(lumA_exp[i,:]) == 0.0:
        rowList_lumA.append(i)

lumA_exp2 = np.zeros(((len(lumA_exp)-len(rowList_lumA)), len(lumA_samples))) # matrix without zero rows
geneList_lumA2 = list()

j = 0 # index in old matrix
r = 0 # index in new matrix
while (j < len(lumA_exp)):
    if j in rowList_lumA:
        j+=1
    else:
        lumA_exp2[r,:] = lumA_exp[j,:]
        geneList_lumA2.append(geneList_lumA[j])
        j+=1
        r+=1
        
# Luminal B data
rowList_lumB = list()
for i in range(0, len(lumB_exp)):
    if np.sum(lumB_exp[i,:]) == 0.0:
        rowList_lumB.append(i)

lumB_exp2 = np.zeros(((len(lumB_exp)-len(rowList_lumB)), len(lumB_samples))) # matrix without zero rows
geneList_lumB2 = list()

j = 0 # index in old matrix
r = 0 # index in new matrix
while (j < len(lumB_exp)):
    if j in rowList_lumB:
        j+=1
    else:
        lumB_exp2[r,:] = lumB_exp[j,:]
        geneList_lumB2.append(geneList_lumB[j])
        j+=1
        r+=1
        
# Basal data
rowList_basal = list()
for i in range(0, len(basal_exp)):
    if np.sum(basal_exp[i,:]) == 0.0:
        rowList_basal.append(i)

basal_exp2 = np.zeros(((len(basal_exp)-len(rowList_basal)), len(basal_samples))) # matrix without zero rows
geneList_basal2 = list()

j = 0 # index in old matrix
r = 0 # index in new matrix
while (j < len(basal_exp)):
    if j in rowList_basal:
        j+=1
    else:
        basal_exp2[r,:] = basal_exp[j,:]
        geneList_basal2.append(geneList_basal[j])
        j+=1
        r+=1
        
# HER2 data
rowList_her2 = list()
for i in range(0, len(her2_exp)):
    if np.sum(her2_exp[i,:]) == 0.0:
        rowList_her2.append(i)

her2_exp2 = np.zeros(((len(her2_exp)-len(rowList_her2)), len(her2_samples))) # matrix without zero rows
geneList_her2_2 = list()

j = 0 # index in old matrix
r = 0 # index in new matrix
while (j < len(her2_exp)):
    if j in rowList_her2:
        j+=1
    else:
        her2_exp2[r,:] = her2_exp[j,:]
        geneList_her2_2.append(geneList_her2[j])
        j+=1
        r+=1

# Normal data
rowList_norm = list()
for i in range(0, len(norm_exp)):
    if np.sum(norm_exp[i,:]) == 0.0:
        rowList_norm.append(i)

norm_exp2 = np.zeros(((len(norm_exp)-len(rowList_norm)), len(norm_samples))) # matrix without zero rows
print(norm_exp2.shape)
geneList_norm2 = list()

print("OLD ", len(norm_exp))
print("ZEROS ", len(rowList_norm))
print("NEW ", len(norm_exp2))

j = 0 # index in old matrix
r = 0 # index in new matrix
while (j < len(norm_exp)):
    if j in rowList_norm:
        j+=1
    else:
        norm_exp2[r,:] = norm_exp[j,:]
        geneList_norm2.append(geneList_norm[j])
        j+=1
        r+=1
        
print(len(geneList_norm2))



## log(TPM + 1) normalization
# TPMi = (FPKMi / sum(FPKMj)) * 10^6, where i is the gene and j is the sample
# logTPM = log(TPM + 1)

# Luminal A
for j in range(0, len(lumA_samples)):
    sumPat = np.sum(lumA_exp2, axis = 0)[j]
    
    for i in range(0, len(lumA_exp2)):
        lumA_exp2[i,j] = np.log2((lumA_exp2[i,j] / sumPat) * (10**6) + 1)
        
# Luminal B
for j in range(0, len(lumB_samples)):
    sumPat = np.sum(lumB_exp2, axis = 0)[j]
    
    for i in range(0, len(lumB_exp2)):
        lumB_exp2[i,j] = np.log2((lumB_exp2[i,j] / sumPat) * (10**6) + 1)
        
# Basal
for j in range(0, len(basal_samples)):
    sumPat = np.sum(basal_exp2, axis = 0)[j]
    
    for i in range(0, len(basal_exp2)):
        basal_exp2[i,j] = np.log2((basal_exp2[i,j] / sumPat) * (10**6) + 1)
        
# HER2
for j in range(0, len(her2_samples)):
    sumPat = np.sum(her2_exp2, axis = 0)[j]
    
    for i in range(0, len(her2_exp2)):
        her2_exp2[i,j] = np.log2((her2_exp2[i,j] / sumPat) * (10**6) + 1)
        
# Normal
for j in range(0, len(norm_samples)):
    sumPat = np.sum(norm_exp2, axis = 0)[j]
    
    for i in range(0, len(norm_exp2)):
        norm_exp2[i,j] = np.log2((norm_exp2[i,j] / sumPat) * (10**6) + 1)

## Extracts TFBM information
# keeps track of which TF and genes have relations as per TFBM file

motifPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/motif.txt' # original motif file

motifTF_lumA = list() # list of TFs interacting to genes that are expressed
motifGenes_lumA = list() # list of genes in expression file with TFs
motifTF_lumB = list()
motifGenes_lumB = list()
motifTF_basal = list()
motifGenes_basal = list()
motifTF_her2 = list()
motifGenes_her2 = list()
motifTF_norm = list()
motifGenes_norm = list()

with open(motifPath, "r") as file:
    for line in file.readlines():
        if line.strip().split("\t")[2] == "1":
            
            # gets motif data for Luminal A
            if line.strip().split("\t")[1] in geneList_lumA2:
                motifTF_lumA.append(line.strip().split("\t")[0])
                motifGenes_lumA.append(line.strip().split("\t")[1])
                
            # gets motif data for Luminal B
            if line.strip().split("\t")[1] in geneList_lumB2:
                motifTF_lumB.append(line.strip().split("\t")[0])
                motifGenes_lumB.append(line.strip().split("\t")[1])
                
            # gets motif data for Basal
            if line.strip().split("\t")[1] in geneList_basal:
                motifTF_basal.append(line.strip().split("\t")[0])
                motifGenes_basal.append(line.strip().split("\t")[1])
                
            # gets motif data for HER2
            if line.strip().split("\t")[1] in geneList_her2:
                motifTF_her2.append(line.strip().split("\t")[0])
                motifGenes_her2.append(line.strip().split("\t")[1])
                
            # gets motif data for Normal
            if line.strip().split("\t")[1] in geneList_norm:
                motifTF_norm.append(line.strip().split("\t")[0])
                motifGenes_norm.append(line.strip().split("\t")[1])

print(len(motifTF_lumA))
print(len(motifGenes_lumA))

print(len(motifTF_lumB))
print(len(motifGenes_lumB))

print(len(motifTF_basal))
print(len(motifGenes_basal))

print(len(motifTF_her2))
print(len(motifGenes_her2))

print(len(motifTF_norm))
print(len(motifGenes_norm))

## Checks if there are genes expressed and not in the motif file
# if not, then doesn't add to updated list of genes in the exp matrix
# and makes note of it to remove later

# Luminal A
removeGene_lumA = list() # list of indeces to remove from expressed genes
for i in range(0, len(geneList_lumA2)):
    if geneList_lumA2[i] not in motifGenes_lumA:
        removeGene_lumA.append(i)
        
print(len(removeGene_lumA))

## Luminal B
removeGene_lumB = list() # list of indeces to remove from expressed genes
for i in range(0, len(geneList_lumB2)):
    if geneList_lumB2[i] not in motifGenes_lumB:
        removeGene_lumB.append(i)
        
print(len(removeGene_lumB))

# Basal
removeGene_basal = list() # list of indeces to remove from expressed genes
for i in range(0, len(geneList_basal2)):
    if geneList_basal2[i] not in motifGenes_basal:
        removeGene_basal.append(i)
        
print(len(removeGene_basal))

# HER2
removeGene_her2 = list() # list of indeces to remove from expressed genes
for i in range(0, len(geneList_her2_2)):
    if geneList_her2_2[i] not in motifGenes_her2:
        removeGene_her2.append(i)
        
print(len(removeGene_her2))

# Normal
removeGene_norm = list() # list of indeces to remove from expressed genes
for i in range(0, len(geneList_norm2)):
    if geneList_norm2[i] not in motifGenes_norm:
        removeGene_norm.append(i)
        
print(len(removeGene_norm))



## Removes the rows and genes that are not in the motif file

# Luminal A
lumA_exp3 = np.zeros((len(lumA_exp2)-len(removeGene_lumA), len(lumA_samples)))
geneList_lumA3 = list() # list of genes that are in motif file and the motifs are expressed genes

r = 0 # index of row in allTesting3
for i in range(0, len(lumA_exp2)):
    if i not in removeGene_lumA:
        lumA_exp3[r,:] = lumA_exp2[i,:]
        geneList_lumA3.append(geneList_lumA2[i])
        r+=1

# Luminal B
lumB_exp3 = np.zeros((len(lumB_exp2)-len(removeGene_lumB), len(lumB_samples)))
geneList_lumB3 = list() # list of genes that are in motif file and the motifs are expressed genes

r = 0 # index of row in allTesting3
for i in range(0, len(lumB_exp2)):
    if i not in removeGene_lumB:
        lumB_exp3[r,:] = lumB_exp2[i,:]
        geneList_lumB3.append(geneList_lumB2[i])
        r+=1

# Basal
basal_exp3 = np.zeros((len(basal_exp2)-len(removeGene_basal), len(basal_samples)))
geneList_basal3 = list() # list of genes that are in motif file and the motifs are expressed genes

r = 0 # index of row in allTesting3
for i in range(0, len(basal_exp2)):
    if i not in removeGene_basal:
        basal_exp3[r,:] = basal_exp2[i,:]
        geneList_basal3.append(geneList_basal2[i])
        r+=1
        
# HER2
her2_exp3 = np.zeros((len(her2_exp2)-len(removeGene_her2), len(her2_samples)))
geneList_her2_3 = list() # list of genes that are in motif file and the motifs are expressed genes

r = 0 # index of row in allTesting3
for i in range(0, len(her2_exp2)):
    if i not in removeGene_her2:
        her2_exp3[r,:] = her2_exp2[i,:]
        geneList_her2_3.append(geneList_her2_2[i])
        r+=1
        
# Normal
norm_exp3 = np.zeros((len(norm_exp2)-len(removeGene_norm), len(norm_samples)))
geneList_norm3 = list() # list of genes that are in motif file and the motifs are expressed genes

r = 0 # index of row in allTesting3
for i in range(0, len(norm_exp2)):
    if i not in removeGene_norm:
        norm_exp3[r,:] = norm_exp2[i,:]
        geneList_norm3.append(geneList_norm2[i])
        r+=1
        
print(norm_exp3.shape)
print(len(geneList_norm3))

## Removes genes the are not expressed but in motif file

# gets all indeces of an element(gene) in a list(motifGenes)
def getIndeces(listOfElements, element):
    indexPosList = []
    indexPos = 0
    while True:
        try:
            # Search for item in list from indexPos to the end of list
            indexPos = listOfElements.index(element, indexPos)
            # Add the index position in list
            indexPosList.append(indexPos)
            indexPos += 1
        except ValueError as e:
            break
 
    return indexPosList

# Luminal A
if set(motifGenes_lumA) != set(geneList_lumA3):
    print(len(set(motifGenes_lumA)))
    print(len(set(geneList_lumA3)))
    
    diff = list(set(motifGenes_lumA) - set(geneList_lumA3))
    print(len(diff))
    for gene in diff:
        geneInds = getIndeces(motifGenes_lumA, gene)

        for i in range(0, len(geneInds)):
            motifGenes_lumA.remove(motifGenes_lumA[geneInds[i] - i])
            motifTF_lumA.remove(motifTF_lumA[geneInds[i] - i])

# Luminal B
if set(motifGenes_lumB) != set(geneList_lumB3):
    print(len(set(motifGenes_lumB)))
    print(len(set(geneList_lumB3)))
    
    diff = list(set(motifGenes_lumB) - set(geneList_lumB3))
    print(len(diff))
    for gene in diff:
        geneInds = getIndeces(motifGenes_lumB, gene)

        for i in range(0, len(geneInds)):
            motifGenes_lumB.remove(motifGenes_lumB[geneInds[i] - i])
            motifTF_lumB.remove(motifTF_lumB[geneInds[i] - i])

# Basal
if set(motifGenes_basal) != set(geneList_basal3):
    print(len(set(motifGenes_basal)))
    print(len(set(geneList_basal3)))
    
    diff = list(set(motifGenes_basal) - set(geneList_basal3))
    print(len(diff))
    for gene in diff:
        geneInds = getIndeces(motifGenes_basal, gene)

        for i in range(0, len(geneInds)):
            motifGenes_basal.remove(motifGenes_basal[geneInds[i] - i])
            motifTF_basal.remove(motifTF_basal[geneInds[i] - i])
            
print(set(motifGenes_basal)== set(geneList_basal3))

# HER2
if set(motifGenes_her2) != set(geneList_her2_3):
    print(len(set(motifGenes_her2)))
    print(len(set(geneList_her2_3)))
    
    diff = list(set(motifGenes_her2) - set(geneList_her2_3))
    print(len(diff))
    for gene in diff:
        geneInds = getIndeces(motifGenes_her2, gene)

        for i in range(0, len(geneInds)):
            motifGenes_her2.remove(motifGenes_her2[geneInds[i] - i])
            motifTF_her2.remove(motifTF_her2[geneInds[i] - i])
            
print(set(motifGenes_her2)== set(geneList_her2_3))

# Normal
if set(motifGenes_norm) != set(geneList_norm3):
    print(len(set(motifGenes_norm)))
    print(len(set(geneList_norm3)))
    
    diff = list(set(motifGenes_norm) - set(geneList_norm3))
    print(len(diff))
    for gene in diff:
        geneInds = getIndeces(motifGenes_norm, gene)

        for i in range(0, len(geneInds)):
            motifGenes_norm.remove(motifGenes_norm[geneInds[i] - i])
            motifTF_norm.remove(motifTF_norm[geneInds[i] - i])
            
print(set(motifGenes_norm)== set(geneList_norm3))



print(len(motifTF_basal))
print(len(motifGenes_basal))

print(len(motifTF_her2))
print(len(motifGenes_her2))

print(len(motifTF_norm))
print(len(motifGenes_norm))

# remove gene rows from HTSeq that are not in the network list

# all_exp_HTSeq is geneList_HTSeq x indivList_HTSeq2

#print(len(all_exp_HTSeq))
#print(len(geneList_HTSeq))
#print(len(indivList_HTSeq2))

network_genes = list(set(geneList_lumA3 + geneList_lumB3 + geneList_basal3 + geneList_her2_3 + geneList_norm3))
print(len(network_genes))

#print(set(geneList_HTSeq))

all_exp_HTSeq2 = np.zeros((len(network_genes), len(indivList_HTSeq2)))
geneList_HTSeq2 = list()

r = 0 # row index of all_exp_HTSeq2
for i in range(0,len(geneList_HTSeq)):
    if geneList_HTSeq[i] in network_genes:
        all_exp_HTSeq2[r,:] = all_exp_HTSeq[i,:]
        geneList_HTSeq2.append(geneList_HTSeq[i])
        r+=1
    
print(len(geneList_HTSeq2))   

## makes the HTSeq file for the differential expression analysis

## Makes the expression file

PANDAexpPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/all_exp_HTSeq2.txt' # change name of new
geneExpFile = open(PANDAexpPath,"w+")

large_string = match_barcodes_HTSeq[0]
for f in range(1, len(match_barcodes_HTSeq)):
    large_string = large_string + "\t" + match_barcodes_HTSeq[f]
    
newLine = large_string + "\n"
geneExpFile.write(newLine)

for i in range(0, len(geneList_HTSeq2)):
    newLine = geneList_HTSeq2[i]
    
    for j in range(0, len(file_id_HTSeq)):
        newLine = newLine + "\t" + str(all_exp_HTSeq2[i,j])
        
    newLine = newLine + "\n"
    geneExpFile.write(newLine)
    
geneExpFile.close()

## Reads in the ppi data and extracts the proteins 
# About 4 hrs to run

ppiPath = '/Users/ursulawidocki/Desktop/BreCaNet/Data/ppi2015_freeze.txt'

protein1_lumA = list()
protein2_lumA = list()
protein1_lumB = list()
protein2_lumB = list()
protein1_basal = list()
protein2_basal = list()
protein1_her2 = list()
protein2_her2 = list()
protein1_norm = list()
protein2_norm = list()

with open(ppiPath, "r") as file:
    next(file) #skips the first line since that is just a header
    for line in file.readlines():
        if (line.strip().split("\t")[0] or line.strip().split("\t")[1]) in motifTF_lumA:
            protein1_lumA.append(line.strip().split("\t")[0])
            protein2_lumA.append(line.strip().split("\t")[1])
        
        if (line.strip().split("\t")[0] or line.strip().split("\t")[1]) in motifTF_lumB:
            protein1_lumB.append(line.strip().split("\t")[0])
            protein2_lumB.append(line.strip().split("\t")[1])
            
        if (line.strip().split("\t")[0] or line.strip().split("\t")[1]) in motifTF_basal:
            protein1_basal.append(line.strip().split("\t")[0])
            protein2_basal.append(line.strip().split("\t")[1])
            
        if (line.strip().split("\t")[0] or line.strip().split("\t")[1]) in motifTF_her2:
            protein1_her2.append(line.strip().split("\t")[0])
            protein2_her2.append(line.strip().split("\t")[1])
            
        if (line.strip().split("\t")[0] or line.strip().split("\t")[1]) in motifTF_norm:
            protein1_norm.append(line.strip().split("\t")[0])
            protein2_norm.append(line.strip().split("\t")[1])

# Check to see if the two protein lists have the same proteins or not

if(set(protein1_lumA) == set(protein2_lumA)):
    print("Same!!")
    
if(set(protein1_lumB) == set(protein2_lumB)):
    print("Same!!")
    
if(set(protein1_basal) == set(protein2_basal)):
    print("Same!!")
    
if(set(protein1_her2) == set(protein2_her2)):
    print("Same!!")
    
if(set(protein1_norm) == set(protein2_norm)):
    print("Same!!")
    
# all proteins sets contain the same proteins

## Makes the expression filess

# Luminal A
PANDAexpPath_lumA = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_lumA.txt'
geneExpFile = open(PANDAexpPath_lumA,"w+")

for i in range(0, len(geneList_lumA3)):
    newLine = geneList_lumA3[i]
    
    for j in range(0, len(lumA_samples)):
        newLine = newLine + "\t" + str(lumA_exp3[i,j])
        
    newLine = newLine + "\n"
    geneExpFile.write(newLine)
    
geneExpFile.close()

# Luminal B
PANDAexpPath_lumB = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_lumB.txt'
geneExpFile = open(PANDAexpPath_lumB,"w+")

for i in range(0, len(geneList_lumB3)):
    newLine = geneList_lumB3[i]
    
    for j in range(0, len(lumB_samples)):
        newLine = newLine + "\t" + str(lumB_exp3[i,j])
        
    newLine = newLine + "\n"
    geneExpFile.write(newLine)
    
geneExpFile.close()

# Basal
PANDAexpPath_basal = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_basal.txt'
geneExpFile = open(PANDAexpPath_basal,"w+")

for i in range(0, len(geneList_basal3)):
    newLine = geneList_basal3[i]
    
    for j in range(0, len(basal_samples)):
        newLine = newLine + "\t" + str(basal_exp3[i,j])
        
    newLine = newLine + "\n"
    geneExpFile.write(newLine)
    
geneExpFile.close()

# HER2
PANDAexpPath_her2 = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_her2.txt'
geneExpFile = open(PANDAexpPath_her2,"w+")

for i in range(0, len(geneList_her2_3)):
    newLine = geneList_her2_3[i]
    
    for j in range(0, len(her2_samples)):
        newLine = newLine + "\t" + str(her2_exp3[i,j])
        
    newLine = newLine + "\n"
    geneExpFile.write(newLine)
    
geneExpFile.close()

# Normal
PANDAexpPath_norm = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_norm.txt'
geneExpFile = open(PANDAexpPath_norm,"w+")

for i in range(0, len(geneList_norm3)):
    newLine = geneList_norm3[i]
    
    for j in range(0, len(norm_samples)):
        newLine = newLine + "\t" + str(norm_exp3[i,j])
        
    newLine = newLine + "\n"
    geneExpFile.write(newLine)
    
geneExpFile.close()

## Makes the motif files

# Luminal A
PANDAmotifPath_lumA = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_lumA.txt'
motifFile = open(PANDAmotifPath_lumA, "w+")

for i in range(0, len(motifTF_lumA)):
    newLine = motifTF_lumA[i] + "\t" + motifGenes_lumA[i] + "\t" + "1.000000" + "\n"
    motifFile.write(newLine)

motifFile.close()

# Luminal B
PANDAmotifPath_lumB = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_lumB.txt'
motifFile = open(PANDAmotifPath_lumB, "w+")

for i in range(0, len(motifTF_lumB)):
    newLine = motifTF_lumB[i] + "\t" + motifGenes_lumB[i] + "\t" + "1.000000" + "\n"
    motifFile.write(newLine)

motifFile.close()

# Basal
PANDAmotifPath_basal = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_basal.txt'
motifFile = open(PANDAmotifPath_basal, "w+")

for i in range(0, len(motifTF_basal)):
    newLine = motifTF_basal[i] + "\t" + motifGenes_basal[i] + "\t" + "1.000000" + "\n"
    motifFile.write(newLine)

motifFile.close()

# HER2
PANDAmotifPath_her2 = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_her2.txt'
motifFile = open(PANDAmotifPath_her2, "w+")

for i in range(0, len(motifTF_her2)):
    newLine = motifTF_her2[i] + "\t" + motifGenes_her2[i] + "\t" + "1.000000" + "\n"
    motifFile.write(newLine)

motifFile.close()

# Normal
PANDAmotifPath_norm = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_norm.txt'
motifFile = open(PANDAmotifPath_norm, "w+")

for i in range(0, len(motifTF_norm)):
    newLine = motifTF_norm[i] + "\t" + motifGenes_norm[i] + "\t" + "1.000000" + "\n"
    motifFile.write(newLine)

motifFile.close()

## Makes the ppi files

# Luminal A
PANDAppiPath_lumA = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_lumA.txt'
ppiFile = open(PANDAppiPath_lumA, "w+")

for i in range(0, len(protein1_lumA)):
    newLine = protein1_lumA[i] + "\t" + protein2_lumA[i] + "\t" + "1" + "\n"
    ppiFile.write(newLine)

ppiFile.close()

# Luminal B
PANDAppiPath_lumB = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_lumB.txt'
ppiFile = open(PANDAppiPath_lumB, "w+")

for i in range(0, len(protein1_lumB)):
    newLine = protein1_lumB[i] + "\t" + protein2_lumB[i] + "\t" + "1" + "\n"
    ppiFile.write(newLine)

ppiFile.close()

# Basal
PANDAppiPath_basal = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_basal.txt'
ppiFile = open(PANDAppiPath_basal, "w+")

for i in range(0, len(protein1_basal)):
    newLine = protein1_basal[i] + "\t" + protein2_basal[i] + "\t" + "1" + "\n"
    ppiFile.write(newLine)

ppiFile.close()

# HER2
PANDAppiPath_her2 = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_her2.txt'
ppiFile = open(PANDAppiPath_her2, "w+")

for i in range(0, len(protein1_her2)):
    newLine = protein1_her2[i] + "\t" + protein2_her2[i] + "\t" + "1" + "\n"
    ppiFile.write(newLine)

ppiFile.close()

# Normal
PANDAppiPath_norm = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_norm.txt'
ppiFile = open(PANDAppiPath_norm, "w+")

for i in range(0, len(protein1_norm)):
    newLine = protein1_norm[i] + "\t" + protein2_norm[i] + "\t" + "1" + "\n"
    ppiFile.write(newLine)

ppiFile.close()


cd Desktop/BreCaNet/Code/pypanda

python run_panda.py -e /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_lumA.txt -m /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_lumA.txt -p /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_lumA.txt -o output_panda_lumA.txt

python run_panda.py -e /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_lumB.txt -m /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_lumB.txt -p /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_lumB.txt -o output_panda_lumB.txt

python run_panda.py -e /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_basal.txt -m /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_basal.txt -p /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_basal.txt -o output_panda_basal.txt

python run_panda.py -e /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_her2.txt -m /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_her2.txt -p /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_her2.txt -o output_panda_her2.txt

python run_panda.py -e /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAgeneExp_norm.txt -m /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAmotifs_norm.txt -p /Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/PANDAppi_norm.txt -o output_panda_norm.txt
