==> BreCaNet/BreCaNet/3_TF_expression.ipynb <==
## Making TF expression

import pandas as pd
import numpy as np


# read in the raw expression data

HTSeq_path = '/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAinput/all_exp_HTSeq.txt'
TF_exp = pd.read_csv(HTSeq_path, delimiter = "\t")
TF_exp.head()

TF_exp.sum(axis = 0)

# read in TFBM

# read in genenames.txt 

# Limitation

not all genes that code for TFs have the same name

######################################

# Old Code

## PANDA output paths
lumA_panda_path = "/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAoutput/output_panda_lumA.txt"
lumB_panda_path = "/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAoutput/output_panda_lumB.txt"
basal_panda_path = "/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAoutput/output_panda_basal.txt"
her2_panda_path = "/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAoutput/output_panda_her2.txt"
norm_panda_path = "/Users/ursulawidocki/Desktop/BreCaNet/Data/PANDAoutput/output_panda_norm.txt"

## Reads in networks

# Luminal A network
G_lumA_df = pd.read_csv(lumA_panda_path, delimiter = " ", names = ["TF", "gene", "prior", "weight"])
TF_lumA = G_lumA_df[["TF", "gene", "prior"]]

# Luminal B network
G_lumB_df = pd.read_csv(lumB_panda_path, delimiter = " ", names = ["TF", "gene", "prior", "weight"])
TF_lumB = G_lumB_df[["TF", "gene", "prior"]]

# Basal network
G_basal_df = pd.read_csv(basal_panda_path, delimiter = " ", names = ["TF", "gene", "prior", "weight"])
TF_basal = G_basal_df[["TF", "gene", "prior"]]

# HER2 network
G_her2_df = pd.read_csv(her2_panda_path, delimiter = " ", names = ["TF", "gene", "prior", "weight"])
TF_her2 = G_her2_df[["TF", "gene", "prior"]]

# Normal network
G_norm_df = pd.read_csv(norm_panda_path, delimiter = " ", names = ["TF", "gene", "prior", "weight"])
TF_norm = G_norm_df[["TF", "gene", "prior"]]



## Initiates TF data frame

column_names = ["TF", "LumA", "LumB", "Basal", "HER2", "Norm"]
TF_df = pd.DataFrame(columns = column_names)
TF_set = set(G_lumA_df["TF"])

TF_df["TF"] = list(TF_set)
TF_df["LumA"] = [0] * TF_df.shape[0]
TF_df["LumB"] = [0] * TF_df.shape[0]
TF_df["Basal"] = [0] * TF_df.shape[0]
TF_df["HER2"] = [0] * TF_df.shape[0]
TF_df["Norm"] = [0] * TF_df.shape[0]
TF_df.index = list(TF_set)
TF_df.head()


def getIndeces(listOfElements, element):
    indexPosList = []
    indexPos = 0
    while True:
        try:
            # Search for item in list from indexPos to the end of list
            indexPos = listOfElements.index(element, indexPos)
            # Add the index position in list
            indexPosList.append(indexPos)
            indexPos += 1
        except ValueError as e:
            break
 
    return indexPosList

## Gets TF prior info and adds to TF_df

for TF in list(TF_df["TF"]):
    indeces = getIndeces(list(TF_lumA["TF"]), TF)
    s = 0
    for ind in indeces:
        s = s + TF_lumA["prior"][ind]
    
    tf = list(TF_df["TF"])
    tf_ind = tf.index(TF)
    TF_df["LumA"][TF] = s

TF_df.head()

for TF in list(TF_df["TF"]):
    indeces = getIndeces(list(TF_lumB["TF"]), TF)
    s = 0
    for ind in indeces:
        s = s + TF_lumB["prior"][ind]
    
    tf = list(TF_df["TF"])
    tf_ind = tf.index(TF)
    TF_df["LumB"][TF] = s

for TF in list(TF_df["TF"]):
    indeces = getIndeces(list(TF_basal["TF"]), TF)
    s = 0
    for ind in indeces:
        s = s + TF_basal["prior"][ind]
    
    tf = list(TF_df["TF"])
    TF_df["Basal"][TF] = s
    
TF_df.head()

for TF in list(TF_df["TF"]):
    indeces = getIndeces(list(TF_her2["TF"]), TF)
    s = 0
    for ind in indeces:
        s = s + TF_her2["prior"][ind]
    
    tf = list(TF_df["TF"])
    TF_df["HER2"][TF] = s
    
TF_df.head()

for TF in list(TF_df["TF"]):
    indeces = getIndeces(list(TF_norm["TF"]), TF)
    s = 0
    for ind in indeces:
        s = s + TF_norm["prior"][ind]
    
    tf = list(TF_df["TF"])
    TF_df["Norm"][TF] = s
    
TF_df.head()

TF_df.to_csv('/Users/ursulawidocki/Desktop/BreCaNet/Data/TF_info.txt', sep='\t')
